cleanup the package directories
