+++ pkg/platform/metrics/constants.go
@@ -1,0 +1,3 @@
+package metrics
+
+// This is for standardized metric constants behaviors that are constants for all the systems.
