+++ pkg/platform/programidentity/constants.go
@@ -1,0 +1,6 @@
+package programidentity
+
+// This package is largely meant to define identity constants that are related to
+// - what is the package constructed by, what's the name of the program
+// - what is the package constructed by, what is the name of the file directory targets
+// etc.
