--- pkg/services/models/wire/wire_test.go
@@ -1,877 +1,0 @@
-package wire
-
-import (
-	"context"
-	"errors"
-	"io"
-	"net/http"
-	"os"
-	"reflect"
-	"runtime"
-	"strings"
-	"testing"
-	"time"
-
-	platformprocess "github.com/portpowered/infinite-you/pkg/platform/process"
-	platformrandom "github.com/portpowered/infinite-you/pkg/platform/random"
-	models "github.com/portpowered/infinite-you/pkg/services/models"
-	modelseffects "github.com/portpowered/infinite-you/pkg/services/models/internal/effects"
-	inference "github.com/portpowered/infinite-you/pkg/services/models/internal/services/inference"
-	"go.uber.org/zap"
-)
-
-func TestNewServiceRejectsMissingConstructionPorts(t *testing.T) {
-	t.Parallel()
-
-	edges := validConstructionEdges()
-	for _, test := range missingConstructionPortCases() {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			assertMissingConstructionPort(t, edges, test)
-		})
-	}
-}
-
-type missingConstructionPortCase struct {
-	name   string
-	mutate func(*constructionEdges)
-	want   string
-}
-
-func missingConstructionPortCases() []missingConstructionPortCase {
-	cases := missingAssetConstructionPortCases()
-	cases = append(cases, missingHostRuntimeConstructionPortCases()...)
-	return cases
-}
-
-func missingAssetConstructionPortCases() []missingConstructionPortCase {
-	return []missingConstructionPortCase{
-		{
-			name: "issuer entropy",
-			mutate: func(edges *constructionEdges) {
-				edges.issuerEntropy = nil
-			},
-			want: "issuer entropy is required",
-		},
-		{
-			name: "asset host platform",
-			mutate: func(edges *constructionEdges) {
-				edges.assetPlatform = models.AssetHostPlatform{}
-			},
-			want: "asset host platform is required",
-		},
-		{
-			name: "asset HTTP client",
-			mutate: func(edges *constructionEdges) {
-				edges.assetHTTP = nil
-			},
-			want: "asset HTTP client is required",
-		},
-		{
-			name: "asset make-directories effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetMkdirAll = nil
-			},
-			want: "asset make-directories effect is required",
-		},
-		{
-			name: "asset inspect-path effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetStat = nil
-			},
-			want: "asset inspect-path effect is required",
-		},
-		{
-			name: "asset resolve-home effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetHome = nil
-			},
-			want: "asset resolve-home effect is required",
-		},
-		{
-			name: "asset write-file effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetWriteFile = nil
-			},
-			want: "asset write-file effect is required",
-		},
-		{
-			name: "asset rename-path effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetRename = nil
-			},
-			want: "asset rename-path effect is required",
-		},
-		{
-			name: "asset remove-path effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetRemove = nil
-			},
-			want: "asset remove-path effect is required",
-		},
-		{
-			name: "asset read-file effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetReadFile = nil
-			},
-			want: "asset read-file effect is required",
-		},
-		{
-			name: "asset read-directory effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetReadDir = nil
-			},
-			want: "asset read-directory effect is required",
-		},
-		{
-			name: "asset create-file effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetCreate = nil
-			},
-			want: "asset create-file effect is required",
-		},
-		{
-			name: "asset open-file effect",
-			mutate: func(edges *constructionEdges) {
-				edges.assetOpen = nil
-			},
-			want: "asset open-file effect is required",
-		},
-	}
-}
-
-func missingHostRuntimeConstructionPortCases() []missingConstructionPortCase {
-	return []missingConstructionPortCase{
-		{
-			name: "process launcher",
-			mutate: func(edges *constructionEdges) {
-				edges.processLauncher = nil
-			},
-			want: "model host process launcher is required",
-		},
-		{
-			name: "host HTTP client",
-			mutate: func(edges *constructionEdges) {
-				edges.hostHTTP = nil
-			},
-			want: "model host HTTP client is required",
-		},
-		{
-			name: "host clock",
-			mutate: func(edges *constructionEdges) {
-				edges.hostClock = nil
-			},
-			want: "model host clock is required",
-		},
-		{
-			name: "runtime command runner",
-			mutate: func(edges *constructionEdges) {
-				edges.runtimeRunner = nil
-			},
-			want: "model runtime command runner is required",
-		},
-		{
-			name: "runtime HTTP client",
-			mutate: func(edges *constructionEdges) {
-				edges.runtimeHTTP = nil
-			},
-			want: "model runtime HTTP client is required",
-		},
-		{
-			name: "runtime file inspector",
-			mutate: func(edges *constructionEdges) {
-				edges.runtimeInspect = nil
-			},
-			want: "model runtime file inspector is required",
-		},
-		{
-			name: "runtime temporary directory resolver",
-			mutate: func(edges *constructionEdges) {
-				edges.runtimeTempDir = nil
-			},
-			want: "model runtime temporary directory resolver is required",
-		},
-		{
-			name: "runtime temporary file creator",
-			mutate: func(edges *constructionEdges) {
-				edges.runtimeTempFile = nil
-			},
-			want: "model runtime temporary file creator is required",
-		},
-		{
-			name: "process clock",
-			mutate: func(edges *constructionEdges) {
-				edges.now = nil
-			},
-			want: "process clock is required",
-		},
-	}
-}
-
-func assertMissingConstructionPort(
-	t *testing.T,
-	edges constructionEdges,
-	test missingConstructionPortCase,
-) {
-	t.Helper()
-
-	current := edges
-	test.mutate(&current)
-	service, err := current.newService()
-	if service != nil {
-		t.Fatal("NewService() returned non-nil service, want nil")
-	}
-	if err == nil || !strings.Contains(err.Error(), test.want) {
-		t.Fatalf("NewService() error = %v, want %q", err, test.want)
-	}
-}
-
-func TestNewServiceReturnsPublishedRootInterface(t *testing.T) {
-	t.Parallel()
-
-	var service models.Service = mustNewWireService(t)
-	if service == nil {
-		t.Fatal("NewService() returned nil service")
-	}
-}
-
-func TestInvocationRuntimeAdaptersPreserveGenericAndASRBackendFacts(t *testing.T) {
-	t.Parallel()
-
-	artifactRef := mustWireTestArtifactRef(t)
-
-	var genericRequest models.InvokeModelRequest
-	genericBackend := wireGenericBackend(artifactRef, &genericRequest)
-	runtime, err := inferenceRuntime(genericBackend, func(context.Context, models.ASRBackendRequest) (models.ASRBackendResponse, error) {
-		return models.ASRBackendResponse{
-			Text:      "transcribed",
-			Segments:  []models.ASRBackendSegment{{ID: 1, Start: 0, End: 10, Text: "transcribed"}},
-			Artifacts: []models.InferenceArtifact{{Artifact: artifactRef, Name: "segments", SizeBytes: 4}},
-		}, nil
-	}, nil)
-	if err != nil {
-		t.Fatalf("inferenceRuntime: %v", err)
-	}
-
-	assertGenericWireInvocation(t, runtime, &genericRequest)
-	assertASRWireInvocation(t, runtime)
-}
-
-func mustWireTestArtifactRef(t *testing.T) models.InferenceArtifactRef {
-	t.Helper()
-	artifactRef, err := (models.InferenceArtifactRef{}).Parse("artifact:wire-test")
-	if err != nil {
-		t.Fatalf("parse artifact reference: %v", err)
-	}
-	return artifactRef
-}
-
-func wireGenericBackend(
-	artifactRef models.InferenceArtifactRef,
-	request *models.InvokeModelRequest,
-) InvocationBackend {
-	return func(_ context.Context, backendRequest models.InvokeModelRequest) ([]models.InferenceContent, []models.InferenceArtifact, error) {
-		*request = backendRequest
-		return []models.InferenceContent{{Name: "text", Content: "generated"}}, []models.InferenceArtifact{{
-			Artifact: artifactRef, Name: "audio", MediaType: "audio/wav", SizeBytes: 12,
-			Properties: map[string]string{"source": "test"},
-		}}, nil
-	}
-}
-
-func assertGenericWireInvocation(t *testing.T, runtime invocationRuntime, request *models.InvokeModelRequest) {
-	t.Helper()
-	generic := models.InvokeModelRequest{ModelName: "voice", Operation: models.OperationTTS}
-	result, err := runtime.Invoke(context.Background(), inference.InvocationRuntimeRequest{Request: generic})
-	if err != nil {
-		t.Fatalf("generic invocation: %v", err)
-	}
-	if !reflect.DeepEqual(*request, generic) || len(result.Content) != 1 || result.Content[0].Content != "generated" {
-		t.Fatalf("generic result = %#v request = %#v, want backend facts and request %#v", result, *request, generic)
-	}
-	if len(result.Artifacts) != 1 || result.Artifacts[0].RefValue != "artifact:wire-test" || result.Artifacts[0].Properties["source"] != "test" {
-		t.Fatalf("generic artifacts = %#v, want detached artifact facts", result.Artifacts)
-	}
-}
-
-func assertASRWireInvocation(t *testing.T, runtime invocationRuntime) {
-	t.Helper()
-	request := models.InvokeModelRequest{
-		Operation: models.OperationASR,
-		Inputs: []models.InferenceInput{
-			{Name: "audio", Modality: models.ModalityAudio, MediaType: "audio/wav", Content: "audio"},
-			{Name: "prompt", Modality: models.ModalityText, ContentType: "text/plain", Content: "meeting"},
-		},
-		Parameters: []models.OperationParameter{{Name: "temperature", Value: 0.2}},
-	}
-	result, err := runtime.Invoke(context.Background(), inference.InvocationRuntimeRequest{
-		Request:   request,
-		Operation: models.Operation{Name: "asr"},
-	})
-	if err != nil {
-		t.Fatalf("ASR invocation: %v", err)
-	}
-	if len(result.Content) != 2 || result.Content[0].Content != "transcribed" || result.Content[1].Name != "segments" {
-		t.Fatalf("ASR content = %#v, want transcript and segments", result.Content)
-	}
-	if len(result.Artifacts) != 1 || result.Artifacts[0].Name != "segments" {
-		t.Fatalf("ASR artifacts = %#v, want mapped backend artifact", result.Artifacts)
-	}
-	if _, err := runtime.Invoke(context.Background(), inference.InvocationRuntimeRequest{Request: request}); err != nil {
-		t.Fatalf("ASR request-operation fallback: %v", err)
-	}
-}
-
-func TestBackendInvocationRuntimePropagatesBackendFailure(t *testing.T) {
-	t.Parallel()
-
-	want := errors.New("backend unavailable")
-	runtime := backendInvocationRuntime{backend: func(context.Context, models.InvokeModelRequest) ([]models.InferenceContent, []models.InferenceArtifact, error) {
-		return nil, nil, want
-	}}
-	if _, err := runtime.Invoke(context.Background(), inference.InvocationRuntimeRequest{}); !errors.Is(err, want) {
-		t.Fatalf("backend invocation error = %v, want %v", err, want)
-	}
-}
-
-func TestNewServiceWithBackendArtifactResolverAndInvocationBackendConstructsRoot(t *testing.T) {
-	t.Parallel()
-
-	service, err := validConstructionEdges().newServiceWithInvocationBackends()
-	if err != nil {
-		t.Fatalf("NewServiceWithBackendArtifactResolverAndInvocationBackend() error = %v", err)
-	}
-	if service == nil {
-		t.Fatal("NewServiceWithBackendArtifactResolverAndInvocationBackend() returned nil service")
-	}
-}
-
-func TestNewInvocationArtifactExporterConstructsPublishedExporter(t *testing.T) {
-	t.Parallel()
-
-	exporter, err := NewInvocationArtifactExporter(invocationArtifactFileSystemStub{})
-	if err != nil {
-		t.Fatalf("NewInvocationArtifactExporter() error = %v", err)
-	}
-	if exporter == nil {
-		t.Fatal("NewInvocationArtifactExporter() returned nil exporter")
-	}
-}
-
-func TestNewServiceHonorsRuntimeAssetEndpointOverrides(t *testing.T) {
-	t.Parallel()
-
-	edges := validConstructionEdges()
-	edges.assetEndpoints = models.RuntimeAssetEndpoints{
-		BaseURL:    "https://example.test/models",
-		APIBaseURL: "https://example.test/api",
-	}
-	service, err := edges.newService()
-	if err != nil {
-		t.Fatalf("NewService() error = %v", err)
-	}
-	if service == nil {
-		t.Fatal("NewService() returned nil service")
-	}
-}
-
-func TestNewServiceWithTypedBackendsConstructsTheAdditiveCompositionPath(t *testing.T) {
-	t.Parallel()
-
-	edges := validConstructionEdges()
-	service, err := NewServiceWithBackendArtifactResolverAndInvocationBackend(
-		edges.assetPlatform,
-		edges.assetHTTP,
-		edges.assetEndpoints,
-		edges.assetMkdirAll,
-		edges.assetStat,
-		edges.assetHome,
-		edges.assetWriteFile,
-		edges.assetRename,
-		edges.assetRemove,
-		edges.assetReadFile,
-		edges.assetReadDir,
-		edges.assetCreate,
-		edges.assetOpen,
-		edges.processLauncher,
-		edges.hostHTTP,
-		edges.hostClock,
-		edges.runtimeRunner,
-		edges.runtimeHTTP,
-		edges.runtimeInspect,
-		edges.runtimeTempDir,
-		edges.runtimeTempFile,
-		zap.NewNop(),
-		edges.now,
-		edges.issuerEntropy,
-		nil,
-		nil,
-		nil,
-		modelseffects.LocalRuntimeHooks{},
-		nil,
-		nil,
-		nil,
-		nil,
-		func(context.Context, models.InvokeModelRequest) ([]models.InferenceContent, []models.InferenceArtifact, error) {
-			return nil, nil, nil
-		},
-		func(context.Context, models.ASRBackendRequest) (models.ASRBackendResponse, error) {
-			return models.ASRBackendResponse{}, nil
-		},
-		func(context.Context, models.EmbeddingBackendRequest) (models.EmbeddingBackendResponse, error) {
-			return models.EmbeddingBackendResponse{}, nil
-		},
-	)
-	if err != nil {
-		t.Fatalf("NewServiceWithBackendArtifactResolverAndInvocationBackend() error = %v", err)
-	}
-	if service == nil {
-		t.Fatal("NewServiceWithBackendArtifactResolverAndInvocationBackend() returned nil service")
-	}
-}
-
-func TestInferenceRuntimeRoutesTypedBackendsAndDetachesResults(t *testing.T) {
-	t.Parallel()
-
-	artifactRef, err := (models.InferenceArtifactRef{}).Parse("artifact:generic")
-	if err != nil {
-		t.Fatalf("parse artifact reference: %v", err)
-	}
-	var genericRequest models.InvokeModelRequest
-	var asrRequest models.ASRBackendRequest
-	var embeddingRequest models.EmbeddingBackendRequest
-	genericBackend := func(_ context.Context, request models.InvokeModelRequest) ([]models.InferenceContent, []models.InferenceArtifact, error) {
-		genericRequest = request
-		return []models.InferenceContent{{ContentType: "text/plain", Content: "generic"}}, []models.InferenceArtifact{{
-			Artifact: artifactRef, Name: "generic", MediaType: "text/plain", SizeBytes: 3,
-			Properties: map[string]string{"kind": "generic"},
-		}}, nil
-	}
-	asrBackend := func(_ context.Context, request models.ASRBackendRequest) (models.ASRBackendResponse, error) {
-		asrRequest = request
-		return models.ASRBackendResponse{
-			Text:     "transcript",
-			Segments: []models.ASRBackendSegment{{ID: 1, Start: 0, End: 2, Text: "transcript"}},
-		}, nil
-	}
-	embeddingBackend := func(_ context.Context, request models.EmbeddingBackendRequest) (models.EmbeddingBackendResponse, error) {
-		embeddingRequest = request
-		return models.EmbeddingBackendResponse{Embeddings: []float64{0.1, 0.2}}, nil
-	}
-
-	runtime, err := inferenceRuntime(genericBackend, asrBackend, embeddingBackend)
-	if err != nil {
-		t.Fatalf("inferenceRuntime() error = %v", err)
-	}
-	assertGenericInferenceRuntimePath(t, runtime, &genericRequest)
-	assertASRInferenceRuntimePath(t, runtime, &asrRequest)
-	assertEmbeddingInferenceRuntimePath(t, runtime, &embeddingRequest)
-	if _, err := genericInvocationRuntime(nil).Invoke(t.Context(), inference.InvocationRuntimeRequest{
-		Request: models.InvokeModelRequest{Input: models.InferenceInput{Content: "echo"}},
-	}); err != nil {
-		t.Fatalf("nil generic backend fallback error = %v", err)
-	}
-}
-
-type ownedCoverageInvocationRuntime interface {
-	Invoke(context.Context, inference.InvocationRuntimeRequest) (inference.InvocationRuntimeResult, error)
-}
-
-func assertGenericInferenceRuntimePath(t *testing.T, runtime ownedCoverageInvocationRuntime, request *models.InvokeModelRequest) {
-	t.Helper()
-	result, err := runtime.Invoke(t.Context(), inference.InvocationRuntimeRequest{
-		Request:   models.InvokeModelRequest{Operation: models.OperationOMNI},
-		Operation: models.Operation{Name: models.OperationOMNI},
-	})
-	if err != nil || len(result.Content) != 1 || result.Content[0].Content != "generic" {
-		t.Fatalf("generic runtime result = %#v, error = %v", result, err)
-	}
-	if len(result.Artifacts) != 1 || result.Artifacts[0].RefValue != "artifact:generic" ||
-		result.Artifacts[0].Properties["kind"] != "generic" {
-		t.Fatalf("generic runtime artifacts = %#v, want detached artifact source", result.Artifacts)
-	}
-	if request.Operation != models.OperationOMNI {
-		t.Fatalf("generic backend request = %#v, want OMNI", request)
-	}
-}
-
-func assertASRInferenceRuntimePath(t *testing.T, runtime ownedCoverageInvocationRuntime, request *models.ASRBackendRequest) {
-	t.Helper()
-	result, err := runtime.Invoke(t.Context(), inference.InvocationRuntimeRequest{
-		Request: models.InvokeModelRequest{
-			Operation: models.OperationASR,
-			Inputs: []models.InferenceInput{
-				{Name: "audio", Modality: models.ModalityAudio, ContentType: "audio/wav", MediaType: "audio/wav", Content: "audio"},
-				{Name: "prompt", Modality: models.ModalityText, ContentType: "text/plain", MediaType: "text/plain", Content: "meeting"},
-			},
-		},
-		Operation: models.Operation{Name: models.OperationASR},
-	})
-	if err != nil || len(result.Content) != 2 || result.Content[0].Name != "transcript" {
-		t.Fatalf("ASR runtime result = %#v, error = %v", result, err)
-	}
-	if string(request.Audio) != "audio" || request.MediaType != "audio/wav" || request.Prompt != "meeting" {
-		t.Fatalf("ASR backend request = %#v, want mapped audio request", request)
-	}
-}
-
-func assertEmbeddingInferenceRuntimePath(t *testing.T, runtime ownedCoverageInvocationRuntime, request *models.EmbeddingBackendRequest) {
-	t.Helper()
-	result, err := runtime.Invoke(t.Context(), inference.InvocationRuntimeRequest{
-		Request: models.InvokeModelRequest{
-			Operation: models.OperationEMBED,
-			Inputs: []models.InferenceInput{
-				{Name: "text", Modality: models.ModalityText, ContentType: "text/plain", MediaType: "text/plain", Content: "find work"},
-				{Name: "parameters", Modality: models.ModalityJSON, ContentType: "application/json", MediaType: "application/json", Content: `{"normalize":true}`},
-			},
-		},
-	})
-	if err != nil || len(result.Content) != 1 || result.Content[0].Name != "embedding" ||
-		result.Content[0].Content != "[0.1,0.2]" {
-		t.Fatalf("embedding runtime result = %#v, error = %v", result, err)
-	}
-	if request.Text != "find work" || request.Parameters["normalize"] != true {
-		t.Fatalf("embedding backend request = %#v, want mapped text and parameter", request)
-	}
-}
-
-func TestInvocationParameterAndArtifactHelpersDetachNestedValues(t *testing.T) {
-	t.Parallel()
-
-	if cloneInvocationParameters(nil) != nil {
-		t.Fatal("cloneInvocationParameters(nil) returned non-nil map")
-	}
-	original := map[string]any{
-		"nested": map[string]any{"value": "original"},
-		"items":  []any{map[string]any{"value": "item"}},
-	}
-	cloned := cloneInvocationParameters(original)
-	original["nested"].(map[string]any)["value"] = "mutated"
-	original["items"].([]any)[0].(map[string]any)["value"] = "mutated"
-	if cloned["nested"].(map[string]any)["value"] != "original" ||
-		cloned["items"].([]any)[0].(map[string]any)["value"] != "item" {
-		t.Fatalf("cloned invocation parameters shared nested values: %#v", cloned)
-	}
-	if invocationArtifactSources(nil) != nil {
-		t.Fatal("invocationArtifactSources(nil) returned non-nil slice")
-	}
-}
-
-type invocationArtifactFileSystemStub struct{}
-
-func (invocationArtifactFileSystemStub) Open(string) (io.ReadCloser, error) {
-	return io.NopCloser(strings.NewReader("")), nil
-}
-
-func (invocationArtifactFileSystemStub) Create(string) (io.WriteCloser, error) {
-	return nopWriteCloser{}, nil
-}
-
-type nopWriteCloser struct{}
-
-func (nopWriteCloser) Write(p []byte) (int, error) { return len(p), nil }
-func (nopWriteCloser) Close() error                { return nil }
-
-func TestNewServiceConstructsInertRoot(t *testing.T) {
-	t.Parallel()
-
-	assetHTTP := &recordingHTTPDoer{name: "asset HTTP"}
-	hostHTTP := &recordingHTTPDoer{name: "host HTTP"}
-	runtimeHTTP := &recordingHTTPDoer{name: "runtime HTTP"}
-	processLauncher := &recordingProcessLauncher{}
-	hostClock := &recordingHostClock{}
-	runtimeRunner := &recordingCommandRunner{}
-	assetMkdirAll := &recordingAssetMkdirAll{}
-	assetStat := &recordingAssetStat{}
-	assetHome := &recordingAssetHome{}
-	assetWriteFile := &recordingAssetWriteFile{}
-	assetRename := &recordingAssetRename{}
-	assetRemove := &recordingAssetRemove{}
-	assetReadFile := &recordingAssetReadFile{}
-	assetReadDir := &recordingAssetReadDir{}
-	assetCreate := &recordingAssetCreate{}
-	assetOpen := &recordingAssetOpen{}
-	runtimeInspect := &recordingRuntimeInspect{}
-	runtimeTempDir := &recordingRuntimeTempDir{}
-	runtimeTempFile := &recordingRuntimeTempFile{}
-	processClock := &recordingProcessClock{}
-
-	runtime.GC()
-	time.Sleep(20 * time.Millisecond)
-	baseline := runtime.NumGoroutine()
-
-	service, err := NewService(
-		models.AssetHostPlatform{OperatingSystem: runtime.GOOS, Architecture: runtime.GOARCH},
-		assetHTTP,
-		models.RuntimeAssetEndpoints{},
-		assetMkdirAll.mkdirAll,
-		assetStat.stat,
-		assetHome.home,
-		assetWriteFile.write,
-		assetRename.rename,
-		assetRemove.remove,
-		assetReadFile.read,
-		assetReadDir.readDir,
-		assetCreate.create,
-		assetOpen.open,
-		processLauncher,
-		hostHTTP,
-		hostClock,
-		runtimeRunner,
-		runtimeHTTP,
-		runtimeInspect.inspect,
-		runtimeTempDir.tempDir,
-		runtimeTempFile.create,
-		zap.NewNop(),
-		processClock.now,
-		platformrandom.CryptoSource{},
-		nil,
-		nil,
-		nil,
-		modelseffects.LocalRuntimeHooks{},
-		nil,
-		nil,
-		nil,
-	)
-	if err != nil {
-		t.Fatalf("NewService() error = %v", err)
-	}
-	if service == nil {
-		t.Fatal("NewService() returned nil service")
-	}
-	assertInertConstruction(t, "asset HTTP", assetHTTP.calls)
-	assertInertConstruction(t, "host HTTP", hostHTTP.calls)
-	assertInertConstruction(t, "runtime HTTP", runtimeHTTP.calls)
-	assertInertConstruction(t, "process launcher", processLauncher.starts)
-	assertInertConstruction(t, "host clock", hostClock.nowCalls+hostClock.timerCalls)
-	assertInertConstruction(t, "runtime command runner", runtimeRunner.calls)
-	assertInertConstruction(t, "asset mkdir", assetMkdirAll.calls)
-	assertInertConstruction(t, "asset stat", assetStat.calls)
-	assertInertConstruction(t, "asset home", assetHome.calls)
-	assertInertConstruction(t, "asset write", assetWriteFile.calls)
-	assertInertConstruction(t, "asset rename", assetRename.calls)
-	assertInertConstruction(t, "asset remove", assetRemove.calls)
-	assertInertConstruction(t, "asset read file", assetReadFile.calls)
-	assertInertConstruction(t, "asset read dir", assetReadDir.calls)
-	assertInertConstruction(t, "asset create", assetCreate.calls)
-	assertInertConstruction(t, "asset open", assetOpen.calls)
-	assertInertConstruction(t, "runtime inspect", runtimeInspect.calls)
-	assertInertConstruction(t, "runtime temp dir", runtimeTempDir.calls)
-	assertInertConstruction(t, "runtime temp file", runtimeTempFile.calls)
-	assertInertConstruction(t, "process clock", processClock.calls)
-
-	runtime.GC()
-	time.Sleep(20 * time.Millisecond)
-	if leaked := runtime.NumGoroutine() - baseline; leaked > 4 {
-		t.Fatalf(
-			"goroutine leak after construction: baseline=%d current=%d delta=%d, want no lifecycle goroutines",
-			baseline, runtime.NumGoroutine(), leaked,
-		)
-	}
-}
-
-func assertInertConstruction(t *testing.T, effect string, calls int) {
-	t.Helper()
-	if calls != 0 {
-		t.Fatalf("construction invoked %s %d times, want inert construction", effect, calls)
-	}
-}
-
-type constructionEdges struct {
-	assetPlatform   models.AssetHostPlatform
-	assetHTTP       modelseffects.AssetHTTPDoer
-	assetEndpoints  models.RuntimeAssetEndpoints
-	assetMkdirAll   modelseffects.AssetMakeDirectories
-	assetStat       modelseffects.AssetInspectPath
-	assetHome       modelseffects.AssetResolveHomeDirectory
-	assetWriteFile  modelseffects.AssetWriteFile
-	assetRename     modelseffects.AssetRenamePath
-	assetRemove     modelseffects.AssetRemovePath
-	assetReadFile   modelseffects.AssetReadFile
-	assetReadDir    modelseffects.AssetReadDirectory
-	assetCreate     modelseffects.AssetCreateFile
-	assetOpen       modelseffects.AssetOpenFile
-	processLauncher modelseffects.HostProcessLauncher
-	hostHTTP        modelseffects.HostHTTPDoer
-	hostClock       modelseffects.HostClock
-	runtimeRunner   platformprocess.CommandRunner
-	runtimeHTTP     modelseffects.RuntimeHTTPDoer
-	runtimeInspect  modelseffects.RuntimeInspectFile
-	runtimeTempDir  modelseffects.RuntimeTempDirectory
-	runtimeTempFile modelseffects.RuntimeCreateTempFile
-	now             func() time.Time
-	issuerEntropy   platformrandom.Source
-}
-
-func validConstructionEdges() constructionEdges {
-	return constructionEdges{
-		assetPlatform: models.AssetHostPlatform{
-			OperatingSystem: runtime.GOOS,
-			Architecture:    runtime.GOARCH,
-		},
-		assetHTTP:       http.DefaultClient,
-		assetMkdirAll:   os.MkdirAll,
-		assetStat:       os.Stat,
-		assetHome:       os.UserHomeDir,
-		assetWriteFile:  os.WriteFile,
-		assetRename:     os.Rename,
-		assetRemove:     os.Remove,
-		assetReadFile:   os.ReadFile,
-		assetReadDir:    os.ReadDir,
-		assetCreate:     func(path string) (io.WriteCloser, error) { return os.Create(path) },
-		assetOpen:       func(path string) (io.ReadCloser, error) { return os.Open(path) },
-		processLauncher: inertProcessLauncher{},
-		hostHTTP:        http.DefaultClient,
-		hostClock:       inertHostClock{},
-		runtimeRunner:   inertCommandRunner{},
-		runtimeHTTP:     http.DefaultClient,
-		runtimeInspect:  os.Stat,
-		runtimeTempDir:  os.TempDir,
-		runtimeTempFile: func(dir, pattern string) (modelseffects.RuntimeTempFile, error) {
-			return os.CreateTemp(dir, pattern)
-		},
-		now:           func() time.Time { return time.Unix(123, 456) },
-		issuerEntropy: platformrandom.CryptoSource{},
-	}
-}
-
-func (edges constructionEdges) newService() (models.Service, error) {
-	return NewServiceWithBackendArtifactResolver(
-		edges.assetPlatform,
-		edges.assetHTTP,
-		edges.assetEndpoints,
-		edges.assetMkdirAll,
-		edges.assetStat,
-		edges.assetHome,
-		edges.assetWriteFile,
-		edges.assetRename,
-		edges.assetRemove,
-		edges.assetReadFile,
-		edges.assetReadDir,
-		edges.assetCreate,
-		edges.assetOpen,
-		edges.processLauncher,
-		edges.hostHTTP,
-		edges.hostClock,
-		edges.runtimeRunner,
-		edges.runtimeHTTP,
-		edges.runtimeInspect,
-		edges.runtimeTempDir,
-		edges.runtimeTempFile,
-		zap.NewNop(),
-		edges.now,
-		edges.issuerEntropy,
-		nil,
-		nil,
-		nil,
-		modelseffects.LocalRuntimeHooks{},
-		nil,
-		nil,
-		nil,
-		nil,
-	)
-}
-
-func (edges constructionEdges) newServiceWithInvocationBackends() (models.Service, error) {
-	return NewServiceWithBackendArtifactResolverAndInvocationBackend(
-		edges.assetPlatform,
-		edges.assetHTTP,
-		edges.assetEndpoints,
-		edges.assetMkdirAll,
-		edges.assetStat,
-		edges.assetHome,
-		edges.assetWriteFile,
-		edges.assetRename,
-		edges.assetRemove,
-		edges.assetReadFile,
-		edges.assetReadDir,
-		edges.assetCreate,
-		edges.assetOpen,
-		edges.processLauncher,
-		edges.hostHTTP,
-		edges.hostClock,
-		edges.runtimeRunner,
-		edges.runtimeHTTP,
-		edges.runtimeInspect,
-		edges.runtimeTempDir,
-		edges.runtimeTempFile,
-		zap.NewNop(),
-		edges.now,
-		edges.issuerEntropy,
-		nil,
-		nil,
-		nil,
-		modelseffects.LocalRuntimeHooks{},
-		nil,
-		nil,
-		nil,
-		nil,
-		nil,
-		nil,
-		nil,
-	)
-}
-
-func mustNewWireService(t *testing.T) models.Service {
-	t.Helper()
-	service, err := validConstructionEdges().newService()
-	if err != nil {
-		t.Fatalf("NewService() error = %v", err)
-	}
-	return service
-}
-
-func TestNewServiceServesPublishedModelsPeerBehavior(t *testing.T) {
-	t.Parallel()
-
-	var service models.Service = mustNewWireService(t)
-	config := models.RuntimeScopeConfig{
-		CacheDirectory: t.TempDir(),
-		Runtime: models.RuntimeConfig{
-			Workers: []models.RuntimeWorker{{
-				Name:          "voice-local",
-				Type:          models.RuntimeWorkerTypeModel,
-				Model:         "OMNIVOICE_Q4_K_M",
-				ModelLocality: models.RuntimeModelLocalityLocal,
-				Operations:    []models.RuntimeOperation{{Name: "TTS"}},
-			}},
-		},
-	}
-	opened, err := service.OpenRuntimeScope(
-		context.Background(),
-		models.OpenRuntimeScopeRequest{Config: config},
-	)
-	if err != nil {
-		t.Fatalf("OpenRuntimeScope: %v", err)
-	}
-
-	readiness, err := service.GetModelReadiness(context.Background(), models.GetModelReadinessRequest{
-		Scope: opened.Scope, Name: "OMNIVOICE_Q4_K_M", Operation: "TTS",
-	})
-	if err != nil {
-		t.Fatalf("GetModelReadiness: %v", err)
-	}
-	if readiness.Readiness.ReadinessState != models.ReadinessStateMissing {
-		t.Fatalf("scoped readiness = %s, want MISSING", readiness.Readiness.ReadinessState)
-	}
-
-	other := mustNewWireService(t)
-	foreign, err := other.OpenRuntimeScope(
-		context.Background(),
-		models.OpenRuntimeScopeRequest{Config: config},
-	)
-	if err != nil {
-		t.Fatalf("open foreign scope: %v", err)
-	}
-	if opened.Scope.String() == foreign.Scope.String() {
-		t.Fatalf("separate Wire-built authorities issued the same scope %q", opened.Scope.String())
-	}
-
-	_, err = service.ListCatalog(
-		context.Background(),
-		models.ListModelsRequest{Scope: foreign.Scope},
-	)
-	if !errors.Is(err, models.ErrRuntimeScopeForeign) {
-		t.Fatalf("ListCatalog foreign scope error = %v, want ErrRuntimeScopeForeign", err)
-	}
-}
