+++ pkg/services/workers/internal/execution/recording/model_test.go
@@ -2,13 +2,10 @@ package recording
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"reflect"
 	"testing"
 	"time"
 
-	workerconfig "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	"github.com/portpowered/infinite-you/pkg/services/providers"
 	"github.com/portpowered/infinite-you/pkg/services/work"
 	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
@@ -18,43 +15,6 @@ type runnerFunc func(context.Context, workerexecution.RunnerExecutionRequest) (w
 
 func (fn runnerFunc) Execute(ctx context.Context, request workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
 	return fn(ctx, request)
-}
-
-func TestRunnerRequiresInjectedClock(t *testing.T) {
-	runner := NewRunner(
-		runnerFunc(func(context.Context, workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-			return workerexecution.RunnerExecutionResult{}, nil
-		}),
-		nil,
-		&workerconfig.FactoryWorkerConfig{Name: "worker"},
-		func(workerexecution.ModelEvent) {},
-		nil,
-	)
-	if _, err := runner.Execute(context.Background(), workerexecution.RunnerExecutionRequest{}); err == nil {
-		t.Fatal("Execute() error = nil, want missing-clock failure")
-	}
-}
-
-func TestModelRecorderPanicDoesNotRewriteRunnerResult(t *testing.T) {
-	runner := NewRunner(
-		runnerFunc(func(context.Context, workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-			return workerexecution.RunnerExecutionResult{Content: "model-result"}, nil
-		}),
-		nil,
-		&workerconfig.FactoryWorkerConfig{Name: "worker", Model: "model"},
-		func(workerexecution.ModelEvent) { panic("recording sink unavailable") },
-		func() time.Time { return time.Unix(600, 0).UTC() },
-	)
-
-	response, err := runner.Execute(context.Background(), workerexecution.RunnerExecutionRequest{
-		Dispatch: work.WorkDispatch{DispatchID: "dispatch-model-sink"},
-	})
-	if err != nil {
-		t.Fatalf("Execute() error = %v, want inner runner result", err)
-	}
-	if response.Content != "model-result" {
-		t.Fatalf("response content = %q, want model-result", response.Content)
-	}
 }
 
 func TestProviderRecorderPanicDoesNotRewriteRunnerResult(t *testing.T) {
@@ -114,151 +74,6 @@ func TestProviderRunnerPreservesCanceledAttemptLineage(t *testing.T) {
 	}
 }
 
-func TestRunnerAndHooksRecordOneCanonicalTrace(t *testing.T) {
-	start := time.Unix(100, 0).UTC()
-	hooks := Hooks()
-	inner := runnerFunc(func(ctx context.Context, _ workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-		hooks.MarkResourceWaitStarted(ctx, start)
-		hooks.MarkResourceWaitFinished(ctx, start.Add(25*time.Millisecond), true)
-		hooks.MarkLoadRequested(ctx, start)
-		hooks.MarkLoadFinished(ctx, start.Add(40*time.Millisecond))
-		hooks.MarkLoadReused(ctx)
-		return workerexecution.RunnerExecutionResult{Content: "done"}, nil
-	})
-
-	var events []workerexecution.ModelEvent
-	times := []time.Time{start, start.Add(100 * time.Millisecond)}
-	now := func() time.Time {
-		value := times[0]
-		times = times[1:]
-		return value
-	}
-	runner := NewRunner(inner, nil, &workerconfig.FactoryWorkerConfig{Name: "worker", Model: "model"}, func(event workerexecution.ModelEvent) {
-		events = append(events, event)
-	}, now)
-
-	_, err := runner.Execute(context.Background(), workerexecution.RunnerExecutionRequest{
-		Dispatch: work.WorkDispatch{DispatchID: "dispatch", Execution: work.ExecutionMetadata{CurrentTick: 3}},
-	})
-	if err != nil {
-		t.Fatalf("Execute: %v", err)
-	}
-	if len(events) != 2 || events[1].Response == nil {
-		t.Fatalf("events = %#v, want request and response", events)
-	}
-	response := events[1].Response
-	if response.ResourceWaitMillis == nil || *response.ResourceWaitMillis != 25 || response.ResourceAcquired == nil || !*response.ResourceAcquired {
-		t.Fatalf("resource trace = %#v, want 25ms acquired", response)
-	}
-	if response.LoadDurationMillis == nil || *response.LoadDurationMillis != 40 || response.LoadRequested == nil || !*response.LoadRequested || response.LoadReused == nil || !*response.LoadReused {
-		t.Fatalf("load trace = %#v, want 40ms requested/reused", response)
-	}
-
-	// Nil contexts are an explicit no-op boundary for model hooks.
-	hooks.MarkResourceWaitStarted(nil, start)
-	hooks.MarkResourceWaitFinished(nil, start, false)
-	hooks.MarkLoadRequested(nil, start)
-	hooks.MarkLoadFinished(nil, start)
-	hooks.MarkLoadReused(nil)
-}
-
-func TestRunnerRecordsDistinctResponseIDsAcrossRetryOutcomes(t *testing.T) {
-	start := time.Unix(200, 0).UTC()
-	calls := 0
-	inner := runnerFunc(func(context.Context, workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-		calls++
-		if calls == 1 {
-			return workerexecution.RunnerExecutionResult{}, errors.New("transient model failure")
-		}
-		return workerexecution.RunnerExecutionResult{Content: "recovered"}, nil
-	})
-
-	var events []workerexecution.ModelEvent
-	times := []time.Time{start, start.Add(time.Millisecond), start.Add(2 * time.Millisecond), start.Add(3 * time.Millisecond)}
-	now := func() time.Time {
-		value := times[0]
-		times = times[1:]
-		return value
-	}
-	runner := NewRunner(inner, nil, &workerconfig.FactoryWorkerConfig{Name: "worker", Model: "model"}, func(event workerexecution.ModelEvent) {
-		events = append(events, event)
-	}, now)
-	request := workerexecution.RunnerExecutionRequest{
-		Dispatch: work.WorkDispatch{DispatchID: "dispatch-retry", Execution: work.ExecutionMetadata{CurrentTick: 4}},
-	}
-
-	executeRetryPair(t, runner, request)
-	assertRetryRecording(t, events)
-}
-
-func TestRunnerResponsePreservesMultipartStructuredContentIncludingAudio(t *testing.T) {
-	want := []work.WorkContentPart{
-		{Type: work.WorkContentPartTypeText, Text: "spoken answer"},
-		{
-			Type:        work.WorkContentPartTypeAudio,
-			URL:         "file:///tmp/spoken-answer.wav",
-			ContentType: "audio/wav",
-			ArtifactID:  "artifact-audio-1",
-			Metadata:    map[string]any{"voice": "alloy"},
-		},
-		{Type: work.WorkContentPartTypeJSON, JSON: json.RawMessage(`{"durationMs":1200}`)},
-	}
-	raw, err := json.Marshal(want)
-	if err != nil {
-		t.Fatalf("marshal structured model output: %v", err)
-	}
-
-	var events []workerexecution.ModelEvent
-	runner := NewRunner(
-		runnerFunc(func(context.Context, workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-			return workerexecution.RunnerExecutionResult{Content: string(raw)}, nil
-		}),
-		nil,
-		&workerconfig.FactoryWorkerConfig{Name: "worker", Model: "model"},
-		func(event workerexecution.ModelEvent) { events = append(events, event) },
-		func() time.Time { return time.Unix(500, 0).UTC() },
-	)
-
-	if _, err := runner.Execute(context.Background(), workerexecution.RunnerExecutionRequest{
-		Dispatch: work.WorkDispatch{DispatchID: "dispatch-structured"},
-	}); err != nil {
-		t.Fatalf("Execute() error = %v", err)
-	}
-	if len(events) != 2 || events[1].Response == nil {
-		t.Fatalf("events = %#v, want one request and one response", events)
-	}
-	if events[1].Response.OutputContent == nil {
-		t.Fatal("MODEL_RESPONSE output content is nil, want the multipart structured result")
-	}
-	if !reflect.DeepEqual(*events[1].Response.OutputContent, want) {
-		t.Fatalf("MODEL_RESPONSE output content = %#v, want %#v", *events[1].Response.OutputContent, want)
-	}
-}
-
-func executeRetryPair(t *testing.T, runner workerexecution.Runner, request workerexecution.RunnerExecutionRequest) {
-	t.Helper()
-	if _, err := runner.Execute(context.Background(), request); err == nil {
-		t.Fatal("first Execute() error = nil, want transient failure")
-	}
-	if _, err := runner.Execute(context.Background(), request); err != nil {
-		t.Fatalf("second Execute() error = %v, want success", err)
-	}
-}
-
-func assertRetryRecording(t *testing.T, events []workerexecution.ModelEvent) {
-	t.Helper()
-	if len(events) != 4 {
-		t.Fatalf("recorded %d events, want request/response pairs for both executions", len(events))
-	}
-
-	firstRequest, firstResponse := events[0], events[1]
-	secondRequest, secondResponse := events[2], events[3]
-	assertRetryResponseIDs(t, firstResponse, secondResponse)
-	wantRequestID := "dispatch-retry/model-request/1"
-	assertRetryRequestIDs(t, firstRequest, secondRequest, wantRequestID)
-	assertRetryResponsePayloads(t, firstResponse, secondResponse, wantRequestID)
-}
-
 func assertRetryResponseIDs(t *testing.T, first, second workerexecution.ModelEvent) {
 	t.Helper()
 	if first.ID != "factory-event/model-response/dispatch-retry/1" {
@@ -302,60 +117,6 @@ func assertRetryResponsePayloads(t *testing.T, first, second workerexecution.Mod
 	}
 	if second.Response.Outcome != workerexecution.InferenceOutcomeSucceeded {
 		t.Fatalf("second response outcome = %q, want succeeded", second.Response.Outcome)
-	}
-}
-
-func TestRunnerResponseOrdinalIsMonotonicAcrossDispatches(t *testing.T) {
-	start := time.Unix(300, 0).UTC()
-	var events []workerexecution.ModelEvent
-	times := []time.Time{
-		start, start.Add(time.Millisecond),
-		start.Add(2 * time.Millisecond), start.Add(3 * time.Millisecond),
-		start.Add(4 * time.Millisecond), start.Add(5 * time.Millisecond),
-	}
-	now := func() time.Time {
-		value := times[0]
-		times = times[1:]
-		return value
-	}
-	runner := NewRunner(
-		runnerFunc(func(context.Context, workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-			return workerexecution.RunnerExecutionResult{Content: "ok"}, nil
-		}),
-		nil,
-		&workerconfig.FactoryWorkerConfig{Name: "worker", Model: "model"},
-		func(event workerexecution.ModelEvent) { events = append(events, event) },
-		now,
-	)
-	for _, dispatchID := range []string{"dispatch-a", "dispatch-b", "dispatch-a"} {
-		if _, err := runner.Execute(context.Background(), workerexecution.RunnerExecutionRequest{
-			Dispatch: work.WorkDispatch{DispatchID: dispatchID},
-		}); err != nil {
-			t.Fatalf("Execute(%q): %v", dispatchID, err)
-		}
-	}
-	if len(events) != 6 {
-		t.Fatalf("recorded %d events, want three request/response pairs", len(events))
-	}
-	got := []string{events[1].ID, events[3].ID, events[5].ID}
-	want := []string{
-		"factory-event/model-response/dispatch-a/1",
-		"factory-event/model-response/dispatch-b/2",
-		"factory-event/model-response/dispatch-a/3",
-	}
-	if !reflect.DeepEqual(got, want) {
-		t.Fatalf("response IDs = %#v, want monotonic runner ordinals %#v", got, want)
-	}
-}
-
-func TestModelResponseEventIDIsDeterministicForAnOrdinal(t *testing.T) {
-	first := modelResponseEventID("dispatch", 2)
-	second := modelResponseEventID("dispatch", 2)
-	if first != second {
-		t.Fatalf("same response ordinal produced IDs %q and %q", first, second)
-	}
-	if first == modelResponseEventID("dispatch", 1) {
-		t.Fatalf("different response ordinals produced the same ID %q", first)
 	}
 }
 
@@ -412,92 +173,5 @@ func TestProviderRunnerRecordsCanonicalEventsAndProviderSession(t *testing.T) {
 	}
 	if events[0].DispatchID != request.Dispatch.DispatchID || events[1].RequestID != request.Dispatch.Execution.RequestID {
 		t.Fatalf("event correlation = %#v/%#v, want dispatch/request IDs", events[0], events[1])
-	}
-}
-
-func TestRunnerFailurePreservesProviderIdentityWithoutSessionID(t *testing.T) {
-	start := time.Unix(450, 0).UTC()
-	var events []workerexecution.ModelEvent
-	runner := NewRunner(
-		runnerFunc(func(context.Context, workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-			return workerexecution.RunnerExecutionResult{}, errors.New("provider unavailable")
-		}),
-		nil,
-		&workerconfig.FactoryWorkerConfig{Name: "worker", Model: "model", ModelProvider: "antigravity"},
-		func(event workerexecution.ModelEvent) { events = append(events, event) },
-		func() time.Time { return start },
-	)
-
-	_, err := runner.Execute(context.Background(), workerexecution.RunnerExecutionRequest{
-		Dispatch:      work.WorkDispatch{DispatchID: "dispatch-provider-failure"},
-		ModelProvider: "antigravity",
-	})
-	if err == nil {
-		t.Fatal("Execute() error = nil, want provider failure")
-	}
-	if len(events) != 2 || events[1].Response == nil {
-		t.Fatalf("events = %#v, want request and response", events)
-	}
-	providerSession := events[1].Response.ProviderSession
-	if providerSession == nil || providerSession.Provider != "antigravity" || providerSession.ID != "" {
-		t.Fatalf("provider session = %#v, want provider-only antigravity metadata", providerSession)
-	}
-}
-
-func TestRunnerRecordsOneStructuredAudioTerminalResponse(t *testing.T) {
-	start := time.Unix(500, 0).UTC()
-	audio := []work.WorkContentPart{{
-		Type:        work.WorkContentPartTypeAudio,
-		File:        "artifacts/response.wav",
-		ContentType: "audio/wav",
-	}}
-	raw, err := json.Marshal(audio)
-	if err != nil {
-		t.Fatalf("marshal audio content: %v", err)
-	}
-
-	var events []workerexecution.ModelEvent
-	times := []time.Time{start, start.Add(4 * time.Millisecond)}
-	now := func() time.Time {
-		value := times[0]
-		times = times[1:]
-		return value
-	}
-	runner := NewRunner(
-		runnerFunc(func(context.Context, workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-			return workerexecution.RunnerExecutionResult{Content: string(raw)}, nil
-		}),
-		nil,
-		&workerconfig.FactoryWorkerConfig{Name: "audio-worker", Model: "audio-model"},
-		func(event workerexecution.ModelEvent) { events = append(events, event) },
-		now,
-	)
-
-	request := workerexecution.RunnerExecutionRequest{
-		Dispatch: work.WorkDispatch{
-			DispatchID: "dispatch-audio",
-			Execution:  work.ExecutionMetadata{RequestID: "request-audio"},
-		},
-	}
-	if _, err := runner.Execute(context.Background(), request); err != nil {
-		t.Fatalf("Execute() error = %v", err)
-	}
-	if len(events) != 2 || events[0].Kind != workerexecution.ModelEventKindRequest ||
-		events[1].Kind != workerexecution.ModelEventKindResponse || events[1].Response == nil {
-		t.Fatalf("events = %#v, want one request and one terminal response", events)
-	}
-	response := events[1].Response
-	if response.Outcome != workerexecution.InferenceOutcomeSucceeded || response.OutputContent == nil {
-		t.Fatalf("response = %#v, want successful structured output", response)
-	}
-	if len(*response.OutputContent) != 1 {
-		t.Fatalf("output content = %#v, want exactly one audio part", *response.OutputContent)
-	}
-	part := (*response.OutputContent)[0]
-	if part.Type != work.WorkContentPartTypeAudio || part.File != audio[0].File || part.ContentType != audio[0].ContentType {
-		t.Fatalf("audio part = %#v, want file/content type preserved", part)
-	}
-	if response.OutputPreview != nil {
-		t.Fatalf("output preview = %q, want nil for structured audio output", *response.OutputPreview)
 	}
 }
