--- tests/functional/internal/support/automations_root.go
@@ -1,28 +1,0 @@
-package support
-
-import (
-	"testing"
-
-	"github.com/portpowered/infinite-you/pkg/root"
-	"github.com/portpowered/infinite-you/pkg/services/automations"
-	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
-)
-
-// AutomationsRootFromProcessEdges returns the published Automations Root composed
-// through the same InjectBundle wiring used by BuildProcess.
-func AutomationsRootFromProcessEdges(
-	t testing.TB,
-	edges serviceedges.Edges,
-	factoryDir string,
-) automations.Root {
-	t.Helper()
-
-	automationsRoot, err := root.AutomationsRootFromEdges(edges, "functional-automations", factoryDir)
-	if err != nil {
-		t.Fatalf("AutomationsRootFromEdges() error = %v", err)
-	}
-	if automationsRoot.Operations == nil {
-		t.Fatal("AutomationsRootFromEdges() returned root without operations")
-	}
-	return automationsRoot
-}
