+++ tests/functional/recordings/root_composition/portable_replay_execution_test.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"io"
 	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strings"
@@ -39,189 +38,6 @@ func TestPortableReplayInspectionExecutesThroughRootProcess(t *testing.T) {
 
 	output := functionalExecutePortableReplay(t, process)
 	functionalAssertPortableReplayInspection(t, output, calls)
-}
-
-// TestPortableReplayResumeProbeUsesRootExecutionOpening proves the
-// checkpoint-bearing replay reaches the canonical Factory Sessions durable
-// owner through the root-composed opening. A public checkpoint summary alone
-// must still produce the non-live replay rejection when no durable snapshot is
-// restorable.
-func TestPortableReplayResumeProbeUsesRootExecutionOpening(t *testing.T) {
-	sessionID := "dur-sess-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-	payload := functionalPortableReplayPayloadForSession(t, sessionID)
-	calls := &functionalReplayLiveConstructionCalls{}
-	home := t.TempDir()
-	projectRoot := t.TempDir()
-	process := support.BuildProcess(t, functionalPortableReplayEdges(t, payload, calls))
-	opening := process.ExecutionRuntimeOpening()
-	if opening == nil {
-		t.Fatal("ExecutionRuntimeOpeningFromProcess() returned nil")
-	}
-	opened, err := opening.OpenExecutionRuntime(t.Context(), factorysessions.ExecutionRuntimeOpeningRequest{
-		ProjectRoot:       projectRoot,
-		SystemConfigHome:  home,
-		FactorySessionID:  sessionID,
-		ReplayPath:        "recording.json",
-		PersistencePolicy: factorysessions.PersistencePolicyDisabled,
-	})
-	if err != nil {
-		t.Fatalf("OpenExecutionRuntime() error = %v", err)
-	}
-	if opened.Close == nil {
-		t.Fatal("OpenExecutionRuntime() close hook = nil")
-	}
-	t.Cleanup(func() {
-		if err := opened.Close(); err != nil {
-			t.Errorf("close replay execution runtime: %v", err)
-		}
-	})
-	if opened.Execution == nil {
-		t.Fatal("OpenExecutionRuntime() execution owner = nil")
-	}
-	if _, err := opened.Execution.ResumeInterruptedSession(
-		t.Context(),
-		sessionID,
-		factorysessions.ResumeSessionRequest{RequestID: "resume-summary-only"},
-	); err == nil {
-		t.Fatal("ResumeInterruptedSession() error = nil, want non-live replay rejection")
-	}
-}
-
-// TestPortableReplayResumeHandoffUsesRootExecutionOpening proves a real
-// interrupted durable snapshot makes the portable replay owner assemble and
-// delegate to the live execution owner. The resumed owner then answers the
-// dispatch read, which keeps the handoff observable at the public durable
-// boundary.
-func TestPortableReplayResumeHandoffUsesRootExecutionOpening(t *testing.T) {
-	const workflowName = "resumable-two-step-fake-children"
-	projectRoot, home, sessionID := functionalInterruptPortableReplay(t, workflowName)
-	functionalResumePortableReplay(t, projectRoot, home, sessionID, workflowName)
-}
-
-func functionalInterruptPortableReplay(t *testing.T, workflowName string) (string, string, string) {
-	projectRoot := functionalWriteResumableWorkflowFixture(t, workflowName)
-	commandRunner := newFunctionalReplayHandoffCommandRunner()
-	home := t.TempDir()
-	env := append(os.Environ(), "HOME="+home, "USERPROFILE="+home)
-	server := support.StartFunctionalAPIServer(t, support.FunctionalAPIServerConfig{
-		FactoryDir: projectRoot,
-		Env:        env,
-		Edges:      serviceedges.Edges{ProviderCommandRunner: commandRunner},
-	})
-	started := functionalPostJSON[factoryapi.FactorySessionExecutionResponse](
-		t,
-		strings.TrimSuffix(server.URL(), "/")+"/factory-sessions/async",
-		factoryapi.FactorySessionExecutionRequest{
-			RequestId: "portable-replay-handoff-start",
-			Source: factoryapi.FactorySessionExecutionSource{
-				Kind:         factoryapi.FactorySessionExecutionSourceKindWorkflowName,
-				WorkflowName: functionalStringPointer(workflowName),
-			},
-			Args: &map[string]any{"subject": "workflows"},
-		},
-	)
-	if started.SessionId == "" {
-		t.Fatal("started session id is empty")
-	}
-	responseStream := support.OpenFactoryResponseEventStreamAt(
-		t,
-		support.SessionResponseEventsURL(server.URL(), started.SessionId),
-	)
-	commandRunner.waitForSecondCall(t)
-	reason := "portable replay handoff interrupt"
-	functionalPostJSON[factoryapi.FactorySessionLifecycleControlResponse](
-		t,
-		strings.TrimSuffix(server.URL(), "/")+"/factory-sessions/"+url.PathEscape(started.SessionId)+"/interrupt-dispatch",
-		factoryapi.FactorySessionInterruptDispatchRequest{DispatchId: "dispatch-2", Reason: &reason},
-	)
-	commandRunner.waitForSecondRelease(t)
-	functionalWaitForInitialResponseTerminal(t, responseStream)
-	responseStream.Close()
-	factoryEvents := support.GetFactoryEventsForSessionAt(t, server.URL(), started.SessionId)
-	interruptedEvent := false
-	for _, event := range factoryEvents {
-		if event.Type == factoryapi.FactoryEventTypeDispatchInterrupted &&
-			event.Context.DispatchId != nil && *event.Context.DispatchId == "dispatch-2" {
-			interruptedEvent = true
-			break
-		}
-	}
-	if !interruptedEvent {
-		t.Fatalf("canonical Factory Events = %#v, want DISPATCH_INTERRUPTED for dispatch-2", factoryEvents)
-	}
-	server.Stop(t)
-	server.Close(t)
-	return projectRoot, home, started.SessionId
-}
-
-func functionalResumePortableReplay(
-	t *testing.T,
-	projectRoot, home, sessionID, workflowName string,
-) {
-	payload := functionalPortableReplayPayloadForSessionSource(t, sessionID, workflowName+".js")
-	edges := functionalPortableReplayEdges(t, payload, &functionalReplayLiveConstructionCalls{})
-	edges.ProviderCommandRunner = support.NewShapedProviderCommandRunner(
-		platformprocess.CommandResult{Stdout: []byte("resumed step")},
-	)
-	process := support.BuildProcess(t, edges)
-	opening := process.ExecutionRuntimeOpening()
-	if opening == nil {
-		t.Fatal("ExecutionRuntimeOpeningFromProcess() returned nil")
-	}
-	opened, err := opening.OpenExecutionRuntime(t.Context(), factorysessions.ExecutionRuntimeOpeningRequest{
-		ProjectRoot:       projectRoot,
-		SystemConfigHome:  home,
-		FactorySessionID:  sessionID,
-		ReplayPath:        "recording.json",
-		PersistencePolicy: factorysessions.PersistencePolicyEnabled,
-	})
-	if err != nil {
-		t.Fatalf("OpenExecutionRuntime(restorable replay) error = %v", err)
-	}
-	if opened.Close == nil || opened.Execution == nil {
-		t.Fatalf("opened restorable runtime = execution:%#v close:%v, want both", opened.Execution, opened.Close != nil)
-	}
-	t.Cleanup(func() {
-		if err := opened.Close(); err != nil {
-			t.Errorf("close restorable replay execution runtime: %v", err)
-		}
-	})
-	resumed, err := opened.Execution.ResumeInterruptedSession(
-		t.Context(),
-		sessionID,
-		factorysessions.ResumeSessionRequest{RequestID: "portable-replay-handoff-resume"},
-	)
-	if err != nil {
-		t.Fatalf("ResumeInterruptedSession(restorable replay) error = %v", err)
-	}
-	if resumed.SessionID != sessionID {
-		t.Fatalf("resumed session id = %q, want %q", resumed.SessionID, sessionID)
-	}
-	responseSubscriber, ok := opened.Execution.(interface {
-		SubscribeResponseEvents(context.Context, string, factorysessions.ResponseEventSubscriptionRequest) (*factorysessions.ResponseEventCursor, error)
-	})
-	if !ok {
-		t.Fatal("resumed execution does not expose response-event subscription")
-	}
-	responseCursor, err := responseSubscriber.SubscribeResponseEvents(
-		t.Context(),
-		sessionID,
-		factorysessions.ResponseEventSubscriptionRequest{SessionID: sessionID},
-	)
-	if err != nil {
-		t.Fatalf("SubscribeResponseEvents(after handoff) error = %v", err)
-	}
-	functionalWaitForResponseTerminal(t, responseCursor)
-	listed, err := opened.Execution.ListDispatches(t.Context(), sessionID)
-	if err != nil {
-		t.Fatalf("ListDispatches(after handoff) error = %v", err)
-	}
-	if len(listed.Dispatches) != 2 {
-		t.Fatalf("ListDispatches(after handoff) count = %d, want two restored dispatches", len(listed.Dispatches))
-	}
-	if err := opened.Close(); err != nil {
-		t.Fatalf("close restorable replay execution runtime: %v", err)
-	}
 }
 
 func functionalWriteResumableWorkflowFixture(t *testing.T, workflowName string) string {
